GigaPath-AI-WSI-Breast-Cancer-Lesion-Analysis/requirements.txt
# GigaPath AI WSI Breast Cancer Lesion Analysis - Dependencies
# Python 3.8+

# ============================================================================
# Core Deep Learning
# ============================================================================
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
scikit-learn>=1.0.0

# ============================================================================
# WSI Processing
# ============================================================================
openslide-python>=1.2.0
Pillow>=9.0.0
opencv-python>=4.7.0

# ============================================================================
# Data Management
# ============================================================================
h5py>=3.7.0
pandas>=1.4.0

# ============================================================================
# Configuration & Utilities
# ============================================================================
PyYAML>=6.0
tqdm>=4.64.0

# ============================================================================
# Visualization & Logging
# ============================================================================
matplotlib>=3.5.0
seaborn>=0.12.0
tensorboard>=2.11.0

# ============================================================================
# Optional: Pre-trained Model Hubs
# ============================================================================
timm>=0.9.0              # PyTorch Image Models (for pretrained backbones)
huggingface-hub>=0.14.0  # For downloading pretrained weights

# ============================================================================
# Development & Testing (Optional)
# ============================================================================
pytest>=7.0.0
jupyter>=1.0.0
ipykernel>=6.0.0
